Car Dealership
Create an automobile class that will be used by a dealership as a vehicle inventory program. The following attributes should be present in your automobile class:

    private string make
    private string model
    private string color
    private int year
    private int mileage

Your program should have appropriate methods such as:

    constructor
    add a new vehicle
    remove a vehicle
    update vehicle attributes

At the end of your program, it should allow the user to output all vehicle inventory to a text file.